generatestatisparam
