+++ next.config.js
@@ -1,6 +1,5 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-    // output: "export",
     images: {
         remotePatterns: [
             {
