+++ src/app/apply/page.js
@@ -4,7 +4,7 @@ import ApplyForm from "@/components/ApplyForm/ApplyForm"
 const page = () => {
   return (
     <>
-    <h1 className="txt-grad leading-loose text-4xl text-center mx-11 font-bold">We Are Hiring</h1>
+    <h1 className="txt-grad leading-loose text-3xl md:text-5xl text-center mx-11  font-extrabold">We Are Hiring !!</h1>
     <ApplyForm />
     </>
   )
