+++ src/app/events/[id]/page.js
@@ -1,6 +1,23 @@
-"use client";
 import EventDetails from "@/components/EventsDetails/EventDetails";
+import { getRequest } from '@/api/api';
 
+export async function generateStaticParams() {
+  try {
+    const response = await getRequest('/api/event/getallevents');
+    const events = await response;
+
+    if (!Array.isArray(events)) {
+      throw new Error('API response is not an array');
+    }
+    
+    return events.map(e => ({
+      id: e._id.toString(),
+    }));
+  } catch (error) {
+    console.error("Failed to fetch events", error);
+    return [];
+  }
+}
 
 const page = ({params}) => {
   return (
