+++ src/app/games/[id]/page.js
@@ -1,7 +1,25 @@
-"use client";
 import React from 'react'
 import GameDetailPage from './GameDetailPage';
 import GameDesignDetails from '@/components/GameDesignDetails';
+import { getRequest } from '@/api/api';
+
+export async function generateStaticParams() {
+  try {
+    const response = await getRequest('/api/game/getallgames');
+    const games = await response;
+
+    if (!Array.isArray(games)) {
+      throw new Error('API response is not an array');
+    }
+    
+    return games.map(e => ({
+      id: e._id.toString(),
+    }));
+  } catch (error) {
+    console.error("Failed to fetch events", error);
+    return [];
+  }
+}
 
 const page = ({params}) => {
   const id=params.id;
