+++ src/components/EventsDetails/EventDetails.js
@@ -1,3 +1,4 @@
+"use client";
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/navigation'
 import eventsData from '../../utilities/EventDetailsData'
